created temp files to add back on adjlist files
