+++ code/adjacencylist_add_on.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include "adjacencylist.h"
+
+/*
+additional data to add for edge struct
+    std::string label;
+    Vertex dest;
+*/
+
+/*//getting vertex, vertices, and the adjacent vertices of selected vertex
+Vertex getVertex(int i);
+std::vector<Vertex> getVertices();
+std::vector<Vertex> getAdjacents(Vertex v);
+
+std::string getLabel(Vertex v);
+void setLabel(Vertex& v, std::string label_str);
+std::string getLabel(Edge e);
+void setLabel(Edge& e, std::string label_str);
+
+std::string getLabel(Vertex v, Vertex w);
+std::string setLabel(Vertex& v, Vertex& w, std::string label_str);*/
